#44 [REFACTOR] utils 파일명 기능 기반으로 변경하여 가독성 개선
