+++ src/utils/auth/find-id/index.ts
@@ -1,2 +1,2 @@
-export * from './validation';
-export * from './format';
+export * from './emailValidation';
+export * from './timerFormat';
