--- src/utils/auth/find-password/format.ts
@@ -1,1 +1,0 @@
-export * from '../find-id/format';
