+++ src/utils/auth/find-password/index.ts
@@ -1,2 +1,2 @@
-export * from './validation';
-export * from './format';
+export * from './passwordValidation';
+export * from '../find-id/timerFormat';
