+++ src/utils/auth/find-password/passwordValidation.ts
