+++ src/utils/chat/messageConverter.ts
