+++ src/utils/common/imageValidator.ts
