+++ src/utils/common/locationFormat.ts
